secured api calls

+++ src/Api/TodoApi.php
@@ -32,30 +32,35 @@ class TodoApi
                 "error" => "[requestType] not defined"
             ];
         } else {
+            //unsecure requestTypes
             switch ($request["requestType"]) {
-                case "dbQuery":
-
-                    $response = $this->Database->apiQuery($request);
-                    break;
-                case "getList":
-                    $response = $this->Database->getList($request);
-                    break;
-                case "removeTask":
-                    $response = $this->Database->removeTask($request);
-                    break;
                 case 'getAuthError':
                     $response = $this->Session->getAuthError();
                     break;
                 default:
-                    $response["msg"] = "given requestType is undefined: \"" . $this->request['requestType'] . "\"";
-                    $response["help"] = "Get help with the api at: https://github.com/Wlczak/gamelist/wiki/Api-backend";
-                    break;
+                    if (!$this->Database->verifyToken()) {
+                        $response['error'] = "Api call authentication failed.";
+                        $response['status'] = false;
+                    } else {
+                        //secure requestTypes
+                        switch ($request["requestType"]) {
+                            case "dbQuery":
+                                $response = $this->Database->apiQuery($request);
+                                break;
+                            case "getList":
+                                $response = $this->Database->getList($request);
+                                break;
+                            case "removeTask":
+                                $response = $this->Database->removeTask($request);
+                                break;
+                            default:
+                                $response["msg"] = "given requestType is undefined: \"" . $this->request['requestType'] . "\"";
+                                $response["help"] = "Get help with the api at: https://github.com/Wlczak/gamelist/wiki/Api-backend";
+                                break;
+                        }
+                    }
             }
         }
-        /*if(){
-
-        }*/
-        //echo gettype($response);
         if (!gettype($response) == "array" || !isset($response)) {
             $response =  ["error" => "Api method error: method returned ivalid values"];
         }
